Restructure vue components
