+++ src/components/common/UserProfileCard.vue
