+++ src/components/editor/InfoForm.vue
