+++ src/components/editor/NoteEditPanel.vue
