+++ src/components/editor/NoteTableItem.vue
