+++ src/components/game/ProgressBar.vue
