+++ src/components/menus/SongListItem.vue
