+++ src/components/ui/FloatingAlert.vue
