+++ src/components/ui/Navbar.vue
