+++ src/mixins/gameMixin.js
