+++ src/routes/Home.vue
@@ -28,7 +28,7 @@
 </template>
 
 <script>
-import Button from "../components/Button.vue";
+import Button from "../components/ui/Button.vue";
 
 export default {
   name: "Home",
